Rename Spirv to SpirvCLI
